+++ packages/ai/src/lib/transforms/undoAI.ts
@@ -1,13 +1,19 @@
 import type { SlateEditor } from 'platejs';
 
+import { getTransientSuggestionKey } from '@platejs/suggestion';
+
 export const undoAI = (editor: SlateEditor) => {
-  if (
-    (editor.history.undos.at(-1) as any)?.ai &&
+  const hasAINodeOrAISuggestion =
     editor.api.some({
       at: [],
       match: (n) => !!(n as any).ai,
-    })
-  ) {
+    }) ||
+    editor.api.some({
+      at: [],
+      match: (n) => !!n[getTransientSuggestionKey()],
+    });
+
+  if ((editor.history.undos.at(-1) as any)?.ai && hasAINodeOrAISuggestion) {
     editor.undo();
     editor.history.redos.pop();
   }