+++ packages/floating/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @platejs/floating
 
+## 50.2.5
+
+### Patch Changes
+
+- [#4644](https://github.com/udecode/plate/pull/4644) by [@felixfeng33](https://github.com/felixfeng33) – Revert #4629
+
+## 50.2.3
+
+### Patch Changes
+
+- [#4629](https://github.com/udecode/plate/pull/4629) by [@narraje](https://github.com/narraje) – Fix: infinite loop in useFloatingToolbar hook
+
 ## 49.0.0
 
 ### Major Changes