+++ packages/selection/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @platejs/selection
 
+## 50.2.0
+
+### Patch Changes
+
+- [#4626](https://github.com/udecode/plate/pull/4626) by [@felixfeng33](https://github.com/felixfeng33) – Improve `insertBlocksAndSelect` to handle fragment insertion better
+
 ## 50.0.0
 
 ### Patch Changes